stt/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
faster-whisper>=1.0.0
python-multipart>=0.0.9
pyyaml>=6.0
